huge push: database now accepts food and exercise removed water model in favor of a single variable in day model renamed day model fixed unit preferences added validation to the email, password at signup

+++ NutritionApp/app/src/main/java/com/example/nutrition/nutritionapp/Model/FoodModel.java
@@ -4,18 +4,16 @@ package com.example.nutrition.nutritionapp.Model;
  * Created by phoenixcampos01 on 9/8/16.
  */
 public class FoodModel {
-    public static final int GRAINS = 0;
-    public static final int DAIRY = 1;
-    public static final int FRUITS = 2;
-    public static final int VEGGIES = 3;
-    public static final int MEAT = 4;
-    public static final int FATS = 5;
-
-
+    public static final double GRAINS = 0.0;
+    public static final double DAIRY = 1.0;
+    public static final double FRUITS = 2.0;
+    public static final double VEGGIES = 3.0;
+    public static final double MEAT = 4.0;
+    public static final double FATS = 5.0;
     private String name;
     private double calories;
     private double servings;
-    //    we need to uniquely identify every food item so we can remove it easily
+    // we need to uniquely identify every food item so we can remove it easily
     private String foodID;
     private double foodType;
 //    Must decide how to split food into the food pyramid
@@ -69,7 +67,7 @@ public class FoodModel {
         return foodType;
     }
 
-    public void setFoodType(int foodType) {
+    public void setFoodType(double foodType) {
         this.foodType = foodType;
     }
